# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Work Stat Lite Api Documentation",
        "description": "Work Stat Lite Api Documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://wslite-backend.herokuapp.com/api//",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/holiday-create": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Holiday create",
                "description": "Holiday create",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "Name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
        "/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get list of all Schedules",
                "description": "Returns list of Schedules",
                "operationId": "GetSchedules",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/schedule-create": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register",
                "description": "Register Schedule",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "Title",
                        "in": "query",
                        "description": "Title",
                        "required": true
                    },
                    {
                        "name": "week_days",
                        "in": "query",
                        "description": "Week Days Json",
                        "required": true
                    },
                    {
                        "name": "schedule_in",
                        "in": "query",
                        "description": "Schedule start time",
                        "required": true
                    },
                    {
                        "name": "schedule_out",
                        "in": "query",
                        "description": "Schedule work time",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/schedule-show": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get schedule",
                "description": "Returns schedule",
                "operationId": "getSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/schedule-update": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Register",
                "description": "Update Schedule",
                "operationId": "postSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    },
                    {
                        "name": "Title",
                        "in": "query",
                        "description": "Title"
                    },
                    {
                        "name": "week_days",
                        "in": "query",
                        "description": "Week Days Json"
                    },
                    {
                        "name": "schedule_in",
                        "in": "query",
                        "description": "Schedule start time"
                    },
                    {
                        "name": "schedule_out",
                        "in": "query",
                        "description": "Schedule work time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule updated successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/schedule-delete": {
            "delete": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete Schedule",
                "description": "Delete Schedul",
                "operationId": "deleteSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/password-reset": {
            "patch": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Update Password Reset records",
                "description": "Update Password Reset records",
                "operationId": "updatePassword Reset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    },
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "Enter old password",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Enter new password"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Enter new password confirmation"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}